requirements1.txt
sentence-transformers
hopsworks
numpy
newsdataapi
python-dotenv
requests
gradio
hopsworks
ipywidgets
newspaper3k
bertopic
nltk
datasets
sklearn
#-U sentence-transformers
twofish
modal
pandas
joblib

--extra-index-url https://download.pytorch.org/whl/cu118
torch 
torchvision 
torchaudio  
